add a bunch more flags and fix code accordingly

+++ src/account.c
@@ -303,5 +303,5 @@ bool account_print_summary(const account_t *acct, int fd) {
   if (len < 0) {
     return false;
   }
-  return write(fd, buffer, len) == len;
+  return write(fd, buffer, (size_t)len) == len;
 }
